requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
python-dotenv
requests
drf-spectacular